requirements.txt
fastapi==0.115.6
python-dotenv==1.0.1
tqdm==4.67.1
openai==1.58.1
pandas==2.2.3
psycopg2-binary==2.9.10
sentence-transformers==3.3.1
uvicorn==0.34.0
pdfplumber==0.11.4
langdetect==1.0.9
langchain==0.1.12
huggingface_hub==0.27.0
dynaconf==3.2.6